Remove incorrect certificate warning
javax.net.ssl.SSLHandshakeException can be thrown for a variety of
reasons that are not related to the certificate being invalid, most
notably socket errors while trying to negotate the TLS session. This can
result in spurious warnings about the remote host serving an invalid
certificate when a connection is closed, and can easily be triggered by
a network change.